+++ ChangeLog.md
@@ -4,6 +4,10 @@
 
 ## Upcoming Release
 
+Table:
+
+- Fix binary table property validation to be 64K bytes not 32K characters
+
 ## 2022.04 Version 3.17.1
 
 Table: